refactor: split cohort selectors into sfc modules
