+++ limen/cohort/__init__.py
@@ -1,15 +1,7 @@
 from limen.cohort.cohort import Cohort
-from limen.cohort.selection import BUILTIN_SELECTORS
-from limen.cohort.selection import select_all
-from limen.cohort.selection import select_backtest_pareto
-from limen.cohort.selection import select_diverse_metrics
-from limen.cohort.selection import select_top_n
+from limen.cohort.sfc import BUILTIN_SELECTORS
 
 __all__ = [
     'BUILTIN_SELECTORS',
     'Cohort',
-    'select_all',
-    'select_backtest_pareto',
-    'select_diverse_metrics',
-    'select_top_n',
 ]
